+++ diplomacy/src/judge/strength.rs
@@ -4,14 +4,14 @@ pub trait Strength {
 }
 
 /// A collection of orders which support a specific order; used in strength calculations.
-pub type Supporters<O> = Vec<O>;
+pub(crate) type Supporters<O> = Vec<O>;
 
 /// The intermediate state for a prevent strength calculation. Prevent strength
 /// determines how much force is applied to stop any other units from entering the
 /// destination province.
 #[derive(Debug, Clone, PartialEq, Eq)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-pub enum Prevent<O> {
+pub(crate) enum Prevent<O> {
     /// The preventing unit cannot reach its destination.
     NoPath,
 
@@ -21,6 +21,16 @@ pub enum Prevent<O> {
 
     /// The order attempts to prevent others from moving to the destination province with support.
     Prevents(O, Supporters<O>),
+}
+
+impl<O: std::fmt::Debug + Copy> Prevent<O> {
+    pub fn unwrap_order(&self) -> O {
+        let Self::Prevents(order, _) = self else {
+            panic!("Attempted to unwrap {:?} value", self);
+        };
+
+        *order
+    }
 }
 
 impl<O> Strength for Prevent<O> {